requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
celery[sqs]==5.3.3
boto3==1.28.40
python-dotenv==1.0.0
pydantic-settings==2.0.3